login functioning and rendering friends to the screen
