+++ friends/src/axios/axios.js
@@ -1,14 +1,14 @@
-import axios from 'axios';
+// import axios from 'axios';
 
-export default function withAuth() {
-  const token = localStorage.getItem('token');
+// export default function axiosWithAuth() {
+//   const token = localStorage.getItem('token');
 
-  const instance = axios.create({
-    headers: {
-      'Content-Type': 'application/json',
-      Authorization: token,
-    },
-  });
+//   const instance = axios.create({
+//     headers: {
+//       'Content-Type': 'application/json',
+//       Authorization: token,
+//     },
+//   });
 
-  return instance;
-}
+//   return instance;
+// }
