+++ friends/src/components/Container.js
@@ -4,13 +4,12 @@ import Login from "./Login";
 import Friends from "./Friends";
 
 export function Container(props) {
-  //   const onLogout = () => {
-  //     // Implement!
-  //     // 1- We need to flush token from local storage
-  //     localStorage.clear();
-  //     // 2- We need to redirect users to login route
-  //     props.history.replace("/");
-  //   };
+  // const onLogout = () => {
+
+  //   localStorage.clear();
+
+  //   props.history.replace("/");
+  // };
 
   return (
     <div className="container">
