+++ friends/src/components/Friends.js
@@ -1,18 +1,35 @@
 import React, { useState, useEffect } from "react";
 import Friend from './Friend';
-import { withAuthCheck } from "./Container";
+import axios from 'axios';
+// import axiosWithAuth from "../axios/axios";
+
+
+
+function axiosWithAuth() {
+  const token = localStorage.getItem('token');
+
+  const instance = axios.create({
+    headers: {
+      'Content-Type': 'application/json',
+      Authorization: token,
+    },
+  });
+
+  return instance;
+}
 
 export default function Friends() {
   const [friends, setFriends] = useState([]);
 
   useEffect(() => {
-    withAuthCheck()
+    axiosWithAuth()
       .get("http://localhost:5000/api/friends")
       .then(res => {
         setFriends(res.data);
+        console.log(friends);
       })
-      .catch(error => {
-        alert(error.response.data.message);
+      .catch(error => {  
+        alert(error.message);
       });
   }, []);
